+++ src/main/java/com/schedule/meet/meeting/repository/SchedulerRepository.java
@@ -12,7 +12,7 @@ import java.util.List;
 public interface SchedulerRepository<P> extends CrudRepository<Scheduler, Long> {
 
     @Query(value = "Select * from scheduler where start_time<:endTime and end_time>:startTime and room_id=:roomId", nativeQuery = true)
-    List<Scheduler> findByTime(@Param("startTime") LocalDateTime startTime, @Param("endTime") LocalDateTime endTime,  @Param("roomId") Integer roomId);
+    List<Scheduler> findByTime(@Param("startTime") LocalDateTime startTime, @Param("endTime") LocalDateTime endTime,  @Param("roomId") Long roomId);
 
 
     @Query(value = "Select * from scheduler where start_time<:endTime and end_time>:startTime", nativeQuery = true)
